+++ NEWS.md
@@ -1,3 +1,8 @@
+# geomultistar 1.1.0
+* Include the [wider] parameter in the [run_geoquery()] function to define layers where geographic elements are not repeated.
+* Function [save_as_geopackage()] to export the results of a geoquery as a layer with an explanatory table of the measurements.
+* Update vignettes and readme.
+
 # geomultistar 1.0.0
 * Initial functionality.
 * Functions to define geographic dimensions: [geomultistar()], [define_geoattribute()] and [get_empty_geoinstances()].